AiModel/requirements_clip.txt
torch
transformers
pillow
opencv-python
requests
python-dotenv
numpy
fastapi
uvicorn[standard]
python-multipart
mediapipe
